03_RAG_Chatbot_System/v1_basic_cli/requirements.txt
langchain_chroma==1.1.0
langchain_classic==1.0.1
langchain_community==0.4.1
langchain_core==1.2.5
langchain_google_genai==4.1.2
langchain_huggingface==1.2.0
langchain_text_splitters==1.1.0
python-dotenv==1.2.1
